requirements.txt
scrapy>=2.11.0
pandas>=2.0.0
typer>=0.9.0
pymongo>=4.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
rich>=13.7.0
lxml>=4.9.0
aiohttp>=3.9.0
colorama>=0.4.6
dnspython>=2.4.0
